requirements.txt
openai
streamlit
faiss-cpu
sentence-transformers
python-dotenv
numpy
